More dropdowns

+++ app/src/main/java/com/browntowndev/liftlab/core/persistence/dtos/MyoRepSetDto.kt
@@ -6,7 +6,7 @@ data class MyoRepSetDto (
     override val id: Long = 0,
     override val workoutLiftId: Long,
     override val position: Int,
-    override val rpeTarget: Double,
+    override val rpeTarget: Float,
     override val repRangeBottom: Int,
     override val repRangeTop: Int,
     val repFloor: Int? = null,
