+++ app/src/main/java/com/browntowndev/liftlab/core/persistence/dtos/StandardSetDto.kt
@@ -6,7 +6,7 @@ data class StandardSetDto (
     override val id: Long = 0,
     override val workoutLiftId: Long,
     override val position: Int,
-    override val rpeTarget: Double,
+    override val rpeTarget: Float,
     override val repRangeBottom: Int,
     override val repRangeTop: Int,
 ) : GenericCustomLiftSet
